% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{type}
\alias{type}
\alias{hash}
\alias{dissect}
\title{Type of Basic Object}
\usage{
type(x)

hash(x)

dissect(x)
}
\arguments{
\item{x}{A Basic object.}
}
\description{
A Basic object can have different internal types. `type` returns the type of
a Basic object. `hash` returns the hash of the object. `dissect` provides some
additional information about the object.
}
