% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{cbind.SymEngineDataType}
\alias{cbind.SymEngineDataType}
\alias{rbind.SymEngineDataType}
\title{Joining DenseMatrix}
\usage{
\method{cbind}{SymEngineDataType}(..., deparse.level)

\method{rbind}{SymEngineDataType}(..., deparse.level)
}
\arguments{
\item{...}{DenseMatrix, VecBasic or R objects.}

\item{deparse.level}{Not used.}
}
\value{
\code{DenseMatrix} S4 object.
}
\description{
S3 methods of \code{cbind} and \code{rbind} defined for
\code{DenseMatrix} and \code{VecBasic}.
}
