% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\title{Code Generation}
\usage{
codegen(x, type = c("ccode", "mathml", "latex", "jscode"))
}
\arguments{
\item{x}{A Basic or a VecBasic object.}

\item{type}{One of "ccode", "mathml", "latex" and "jscode".}
}
\value{
A character vector.
}
\description{
Generate C/MathML/LaTeX/JavaScript code string from a \code{Basic}
or \code{VecBasic} object.
}
