% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{as.character,Basic-method}
\alias{as.character,Basic-method}
\alias{as.numeric,Basic-method}
\alias{as.integer,Basic-method}
\title{Some Conversion Methods}
\usage{
\S4method{as.character}{Basic}(x)

\S4method{as.numeric}{Basic}(x)

\S4method{as.integer}{Basic}(x)
}
\arguments{
\item{x}{The object to be converted.}
}
\description{
Some Conversion Methods
}
