% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{D}
\alias{D}
\title{Derivatives of Basic Objects}
\usage{
D(expr, name, n = 1L)
}
\arguments{
\item{expr}{A Basic object.}

\item{name}{A character vector or a Basic object of type Symbol.}

\item{n}{An integer representing the level of derivative.}
}
\description{
Derivatives of Basic Objects
}
