% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{as.matrix.DenseMatrix}
\alias{as.matrix.DenseMatrix}
\alias{dim,DenseMatrix-method}
\alias{dim<-,DenseMatrix-method}
\alias{dim<-,VecBasic-method}
\alias{dim<-,Basic-method}
\alias{dimnames<-,DenseMatrix-method}
\alias{dimnames,DenseMatrix-method}
\alias{length,DenseMatrix-method}
\alias{[[,DenseMatrix,ANY-method}
\alias{[[<-,DenseMatrix-method}
\alias{[,DenseMatrix-method}
\alias{[<-,DenseMatrix-method}
\title{Methods Related to DenseMatrix}
\usage{
\method{as.matrix}{DenseMatrix}(x, ...)

\S4method{dim}{DenseMatrix}(x)

\S4method{dim}{DenseMatrix}(x) <- value

\S4method{dim}{VecBasic}(x) <- value

\S4method{dim}{Basic}(x) <- value

\S4method{dimnames}{DenseMatrix}(x) <- value

\S4method{dimnames}{DenseMatrix}(x)

\S4method{length}{DenseMatrix}(x)

\S4method{[[}{DenseMatrix,ANY}(x, i, j, ...)

\S4method{[[}{DenseMatrix}(x, i, j, ...) <- value

\S4method{[}{DenseMatrix}(x, i, j, ..., drop = TRUE)

\S4method{[}{DenseMatrix}(x, i, j, ...) <- value
}
\arguments{
\item{x}{A DenseMatrix object.}

\item{i, j, value, ..., drop}{Arguments for subsetting, assignment or replacing.}
}
\value{
Same or similar with the generics of these methods.
}
\description{
These are miscellaneous S3/S4 methods defined for \code{DenseMatrix} class.
}
