% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_symbol.R
\name{Function}
\alias{Function}
\alias{FunctionSymbol}
\title{Create a FunctionSymbol}
\usage{
Function(name)

FunctionSymbol(name, args)
}
\arguments{
\item{name}{Name of the function symbol}

\item{args}{Dependent symbols}
}
\description{
\code{FunctionSymbol} creates a Basic object with type \code{FunctionSymbol}.
\code{Function} returns a generator.
}
