% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdify.R
\name{lambdify}
\alias{lambdify}
\title{Convert Basic Object to R Function}
\usage{
lambdify(x, backend = c("auto", "lambda", "llvm"), perform_cse = TRUE)
}
\arguments{
\item{x}{A Basic object or a VecBasic object.}

\item{backend}{One of "auto", "lambda" and "llvm". If "auto", 
\code{getOption("lambdify.backend")} will be used to determine the value. If that
option is not set, it will be determined based on \code{symengine_have_component("llvm")}.}

\item{perform_cse}{Passed to \code{\link{DoubleVisitor}}.}
}
\description{
Convert Basic Object to R Function
}
