% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{LCM}
\alias{LCM}
\alias{GCD}
\alias{nextprime}
\alias{factorial}
\alias{factorial,SymEngineDataType-method}
\alias{choose}
\alias{choose,SymEngineDataType-method}
\alias{zeta}
\alias{lambertw}
\alias{dirichlet_eta}
\alias{erf}
\alias{erfc}
\title{Some Special Math Functions}
\usage{
LCM(a, b)

GCD(a, b)

nextprime(a)

factorial(x)

\S4method{factorial}{SymEngineDataType}(x)

choose(n, k)

\S4method{choose}{SymEngineDataType}(n, k)

zeta(a)

lambertw(a)

dirichlet_eta(a)

erf(a)

erfc(a)
}
\arguments{
\item{a, b, x, n, k}{SymEngine objects. Some functions require Integer type.}
}
\description{
Some special mathematical functions and functions related to number theory.
}
