% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{Matrix}
\alias{Matrix}
\title{DenseMatrix Constructor}
\usage{
Matrix(data, nrow = 1L, ncol = 1L, byrow = FALSE)
}
\arguments{
\item{data}{A R object.}

\item{nrow, ncol}{Number of rows and columns.}

\item{byrow}{Boolean value. Whether the data should be filled by row or by column.}
}
\description{
DenseMatrix Constructor
}
