% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{s4basic_parse}
\alias{s4basic_parse}
\title{Main entry point for converting scalar R object to Basic}
\usage{
s4basic_parse(robj, check_whole_number = FALSE)
}
\arguments{
\item{robj}{A R object to convert.}

\item{check_whole_number}{Boolean value indicating whehter to check a
double input is a whole number. If TRUE, it will be converted to
Integer.}
}
\description{
Main entry point for converting scalar R object to Basic
}
\examples{
symengine:::s4basic_parse("x")
symengine:::s4basic_parse(1, check_whole_number = TRUE) ==
symengine:::s4basic_parse(1L, check_whole_number = TRUE)
symengine:::s4basic_parse(NA_real_)
}
