% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{methods}
\name{solve}
\alias{solve}
\alias{solve,DenseMatrix-method}
\alias{solve,VecBasic-method}
\alias{solve,Basic-method}
\title{Solve Symbolic Equations}
\usage{
solve(a, b, ...)

\S4method{solve}{DenseMatrix}(a, b, ...)

\S4method{solve}{VecBasic}(a, b, ...)

\S4method{solve}{Basic}(a, b, ...)
}
\arguments{
\item{a, b}{Objects, see details.}

\item{...}{Not used.}
}
\description{
Solve system of equations or a polynomial equation.
}
\details{
\code{solve} is a generic function dispatched on the class of the first argument.
If \code{a} is a (square) DenseMatrix, it solves the equation \code{a \%*\% x = b}.
If \code{a} is a DenseMatrix and \code{b} is missing, \code{b} is taken to be an
identity matrix and \code{solve} will return the inverse of \code{a}.
If \code{a} is a VecBasic, it solves the system of linear equations represented by
\code{a} with regards to symbols represented in \code{b}.
If \code{a} is a Basic, it solves the polynomial equation represented by a with regards
to the symbol represented in \code{b}.
}
