% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{subs}
\alias{subs}
\title{Substitute Expressions in SymEngine Objects}
\usage{
subs(expr, ...)
}
\arguments{
\item{expr}{A \code{Basic} S4 object.}

\item{...}{Pairs of Basic objects or values can be converted to \code{Basic}.
In the order of "from1, to1, from2, to2, ...".}
}
\value{
Same type as \code{expr}.
}
\description{
This function will substitute \code{expr} with pairs of
values in the dot arguments. The length of dot arguments must
be a even number.
}
