% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symengine_info.R
\name{symengine_version}
\alias{symengine_version}
\alias{symengine_ascii_art}
\alias{symengine_have_component}
\alias{symengine_compilation_notes}
\title{Information about SymEngine Library}
\usage{
symengine_version()

symengine_ascii_art()

symengine_have_component(which = c("mpfr", "flint", "arb", "mpc", "ecm",
  "primesieve", "piranha", "boost", "pthread", "llvm"))

symengine_compilation_notes()
}
\description{
Functions to get symengine logo, version and external libraries built with.
}
