% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{t}
\alias{t}
\alias{t,Basic-method}
\alias{t,VecBasic-method}
\alias{t,DenseMatrix-method}
\title{Transpose (as) a DenseMatrix}
\usage{
t(x)

\S4method{t}{Basic}(x)

\S4method{t}{VecBasic}(x)

\S4method{t}{DenseMatrix}(x)
}
\arguments{
\item{x}{A SymEngine object.}
}
\value{
A \code{DenseMatrix} S4 object.
}
\description{
S4 methods of \code{t} defined for \code{Basic}, \code{VecBasic}
and \code{DenseMatrix}.
}
