% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{length,VecBasic-method}
\alias{length,VecBasic-method}
\alias{rep,VecBasic-method}
\alias{rep,Basic-method}
\alias{unique.VecBasic}
\alias{c,BasicOrVecBasic-method}
\alias{[[,VecBasic,numeric-method}
\alias{[,VecBasic-method}
\alias{[[<-,VecBasic-method}
\alias{[<-,VecBasic-method}
\title{Methods Related to VecBasic}
\usage{
\S4method{length}{VecBasic}(x)

\S4method{rep}{VecBasic}(x, ...)

\S4method{rep}{Basic}(x, ...)

\method{unique}{VecBasic}(x, ...)

\S4method{c}{BasicOrVecBasic}(x, ...)

\S4method{[[}{VecBasic,numeric}(x, i, j, ...)

\S4method{[}{VecBasic}(x, i, j, ..., drop = TRUE)

\S4method{[[}{VecBasic}(x, i) <- value

\S4method{[}{VecBasic}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Basic object or Vecbasic object.}

\item{i, j, ..., drop, value}{Arguments for subsetting or replacing.}
}
\value{
Same or similar to the generics.
}
\description{
Miscellaneous S4 methods defined for \code{VecBasic} class.
}
