% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{Vector}
\alias{Vector}
\alias{V}
\title{VecBasic Constructors}
\usage{
Vector(x, ...)

V(...)
}
\arguments{
\item{x, ...}{R objects.}
}
\description{
VecBasic Constructors
}
\examples{
a <- S("a")
b <- S("b")
Vector(a, b, a + b, 42L)
Vector(list(a, b, 42L))

Vector(1,2,a)
V(1,2,a)
}
