% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-getinfo.R
\name{get_type}
\alias{get_type}
\alias{get_args}
\alias{get_hash}
\alias{get_str}
\alias{free_symbols}
\alias{function_symbols}
\alias{get_name}
\alias{get_prec}
\title{Get Information about Basic Object}
\usage{
get_type(x)

get_args(x)

get_hash(x)

get_str(x)

free_symbols(x)

function_symbols(x)

get_name(x)

get_prec(x)
}
\arguments{
\item{x}{A Basic object.}
}
\value{
\itemize{
    \item \code{get_type()}, \code{get_hash()}, \code{get_str()}, \code{get_name()}
          return a string.
    \item \code{get_args()}, \code{free_symbols()}, \code{function_symbols()}
          return a \code{VecBasic} S4 object.
    \item \code{get_prec()} returns an integer.
  }
}
\description{
These functions are used to access the underlying properties of a
\code{Basic} object.
}
\details{
\describe{
  \item{get_type}{Return the internal type}
  \item{get_args}{Return the internal arguments of a Basic object as a VecBasic}
  \item{get_hash}{Return the hash as a string}
  \item{get_str}{Return the string representation of the Basic object}
  \item{free_symbols}{Return free symbols in an expression}
  \item{function_symbols}{Return function symbols in an expression}
  \item{get_name}{Return name of a Basic object of type FunctionSymbol}
  \item{get_prec}{Return precision of a Basic object of type RealMPFR}
}
}
