% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{codegen}
\alias{codegen}
\title{Generate C/mathml/latex/JavaScript String}
\usage{
codegen(x, type = c("ccode", "mathml", "latex", "jscode"))
}
\arguments{
\item{x}{A Basic or a VecBasic object.}

\item{type}{One of "ccode", "mathml", "latex" and "jscode".}
}
\description{
Generate C/mathml/latex/JavaScript String
}
