% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{as.character,Basic-method}
\alias{as.character,Basic-method}
\alias{as.numeric,Basic-method}
\alias{as.integer,Basic-method}
\alias{as.character,VecBasic-method}
\alias{as.numeric,VecBasic-method}
\alias{as.integer,VecBasic-method}
\alias{as.language}
\alias{as.language,Basic-method}
\title{Some Conversion Methods}
\usage{
\S4method{as.character}{Basic}(x)

\S4method{as.numeric}{Basic}(x)

\S4method{as.integer}{Basic}(x)

\S4method{as.character}{VecBasic}(x)

\S4method{as.numeric}{VecBasic}(x)

\S4method{as.integer}{VecBasic}(x)

as.language(x)

\S4method{as.language}{Basic}(x)
}
\arguments{
\item{x}{The object to be converted.}
}
\value{
Same as default methods of these generics. \code{as.language()}
  may return \code{symbol}, \code{integer}, \code{double} or \code{call}.
}
\description{
Miscellaneous S4 methods defined for converting a \code{Basic} or
\code{VecBasic} object to R number/string/language object.
}
