% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{D,SymEngineDataType-method}
\alias{D,SymEngineDataType-method}
\title{Derivatives of Basic Objects}
\usage{
\S4method{D}{SymEngineDataType}(expr, name)
}
\arguments{
\item{expr}{A Basic object.}

\item{name}{A character vector or a Basic object of type Symbol.}
}
\description{
Derivatives of Basic Objects
}
