% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{det}
\alias{det}
\alias{det,DenseMatrix-method}
\title{Calculate the Determinant of DenseMatrix}
\usage{
det(x, ...)

\S4method{det}{DenseMatrix}(x, ...)
}
\arguments{
\item{x}{A DenseMatrix object.}

\item{...}{Unused.}
}
\value{
A \code{Basic} object.
}
\description{
S4 method of \code{det} defined for \code{DenseMatrix}.
}
\examples{
mat <- Matrix(LETTERS[1:9], 3)
det(mat)
}
