% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_symbol.R
\name{Function}
\alias{Function}
\alias{FunctionSymbol}
\title{Create a FunctionSymbol}
\usage{
Function(name)

FunctionSymbol(name, args)
}
\arguments{
\item{name}{Name of the function symbol}

\item{args}{Dependent symbols}
}
\value{
\code{FunctionSymbol} returns a \code{Basic}. \code{Function} returns
  a function that will return a \code{Basic}
}
\description{
\code{FunctionSymbol} creates a Basic object with type \code{FunctionSymbol}.
\code{Function} returns a generator.
}
\examples{
f <- Function("f")
a <- Symbol("a")
b <- Symbol("b")
f(a, b)
e <- f(a, f(a + b))
D(e, a)
FunctionSymbol("f", c(a,b))
}
\seealso{
\code{\link{S}}
}
