% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{LCM}
\alias{LCM}
\alias{GCD}
\alias{nextprime}
\alias{factorial}
\alias{factorial,SymEngineDataType-method}
\alias{choose}
\alias{choose,SymEngineDataType-method}
\alias{zeta}
\alias{lambertw}
\alias{dirichlet_eta}
\alias{erf}
\alias{erfc}
\alias{atan2,SymEngineDataType,SymEngineDataType-method}
\alias{kronecker_delta}
\alias{lowergamma}
\alias{uppergamma}
\alias{beta,SymEngineDataType,SymEngineDataType-method}
\alias{psigamma,SymEngineDataType-method}
\alias{digamma,SymEngineDataType-method}
\alias{trigamma,SymEngineDataType-method}
\title{Some Special Math Functions}
\usage{
LCM(a, b)

GCD(a, b)

nextprime(a)

factorial(x)

\S4method{factorial}{SymEngineDataType}(x)

choose(n, k)

\S4method{choose}{SymEngineDataType}(n, k)

zeta(a)

lambertw(a)

dirichlet_eta(a)

erf(a)

erfc(a)

\S4method{atan2}{SymEngineDataType,SymEngineDataType}(y, x)

kronecker_delta(x, y)

lowergamma(x, a)

uppergamma(x, a)

\S4method{beta}{SymEngineDataType,SymEngineDataType}(a, b)

\S4method{psigamma}{SymEngineDataType}(x, deriv = 0L)

\S4method{digamma}{SymEngineDataType}(x)

\S4method{trigamma}{SymEngineDataType}(x)
}
\arguments{
\item{a, b, x, y, n, k, deriv}{SymEngine objects (\code{Basic}/\code{VecBasic}/\code{DenseMatrix}).
Some functions require Integer type.}
}
\value{
Same type as input.
}
\description{
These are some special mathematical functions and functions
related to number theory.
}
