% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\docType{methods}
\name{solve}
\alias{solve}
\alias{solve,DenseMatrix-method}
\title{Solve Symbolic Equations}
\usage{
solve(a, b, ...)

\S4method{solve}{DenseMatrix}(a, b, ...)
}
\arguments{
\item{a}{A square DenseMatrix object.}

\item{b}{A VecBasic or a DenseMatrix giving the right hand side of
the equation. If missing, b is taken to be an identity matrix
and `solve` will return the inverse of `a`.}

\item{...}{Not used.}
}
\description{
`solve` solves the equation "a %*% x = b".
}
