% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{subs}
\alias{subs}
\title{Substitute Expressions in SymEngine Objects}
\usage{
subs(expr, ...)
}
\arguments{
\item{expr}{A SymEngine object.}

\item{...}{Pairs of Basic objects, e.g. (from1, to1, from2, to2).}
}
\description{
Substitute Expressions in SymEngine Objects
}
